backend/requirements.txt
# Core FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL driver
# sqlite support is built-in

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# AI & Computer Vision
face-recognition==1.3.0  # Simple face recognition library
# OR use these alternatives:
# deepface==0.0.79
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
dlib==19.24.2  # Required for face_recognition

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0  # For testing async endpoints

# Optional: Redis for caching
# redis==5.0.1

# Optional: File storage
# boto3==1.34.30  # AWS S3
